Fixed layout packing for MSVC compilers
As opposed to GCC and GCC-like compilers, MSVC always defaults to using
32-bit when using an enum as a struct member, even for packed ones and
even if the enum values are explicitly defined as 8 or 16 bit.

This obviously leads to the layout structs being incorrectly packed and
the _Static_assert() struct size validation being triggered.

Fix this by using explicit size-typed members in the layout structs.