Replaced va_### variable arg types with ntfs_va_###
Because UEFI toolchains may not define the UEFI VA_### types to something
that can be mapped to the (sometimes builtin) va_### types used by some
toolchains, we have to ensure that the va_### definitions used in logging.c
match the VA_### ones from UEFI.

We therefore introduce the ntfs_va_### types that map to the UEFi VA_###
types, when compiling the UEFI driver, and map to the compiler's original
va_### types otherwise. We also define the VA_### types for gnu-efi.