Added UEFI driver FileOpen/FileClose/FileDelete
Note that, because of the nature of the UEFI environments, where users
may literally yank a power cable to power down their system while using
the UEFI Shell (e.g. Raspberry Pi), we have to ensure that NTFS volumes
remain mounted for as little time as possible so that we can avoid the
dreaded "Unclean NTFS volume" on the next remount.

To that effect, we track the number of all files open on a volume and
unmount that volume, in FileClose(), if that number reaches 0 (since
this should only happen when the root file handle that was returned on
OpenVolume is being closed).

Obviously, this repeated mounting and unmounting of the NTFS volume
will have a detrimental effect on performance however...