Fixed --enable-debug configure option
The --enable-debug option sets ENABLE_DEBUG in config.h but the sources all
appear to rely on DEBUG being defined instead, which makes the option useless.
Also the use of DEBUG as a macro variable conflicts with the DEBUG() macro
function, that is defined by EDK2 and gnu-efi when compiling UEFI executables.

We therefore fix all these issues by replacing all instances of the DEBUG
macro with ENABLE_DEBUG.