Switched readdir to return -1 for pos on EOD condition
Using a directory specific position for the EOD condition is not really
helpful unless you know the NTFS internals of a specific directory, which
is no expected to be the case for consumers of libntfs-3g.

Instead, switch to always returning a fixed pos value of -1 to mark EOD
so that applications that may call into readdir repeatedly and want to
know if the EOD condition has been reached can easily find that out.

Also clarify code and fix some missing line feed in the log messages.