Clarified readdir code and switched to position based loops
Removed the position computation from ntfs_filldir() to perform it in the
loops instead. This aims at facilitating adding the capability to restart
a directory listing from a specific position.

Also introduced a new ia_offset variable, that simplifies pos computation
and helps with pos to index entry address translation, which is now moved
to the beginning of each loop.