Updated readdir to drop the use of the bmp_buf_pos variable
There's little benefit in issuing partial reads of the bitmap table NTFS
attribute, so that we can save a few bytes, when the performance impact of
always readding it once, in full, is negligible and doing so helps simplify
the bitmap lookup code by using a single bitmap position index. Furthermore,
trying to reuse a bitmap positioning code with two unsynchronized indexes
would add a lot of complexity when trying to continue a directory read from
a known position (because we can't easily derive bmp_buf_pos from buf_pos).

So switch to reading the full bitmap table always (which is capped at 4 KB
anyway, and therefore won't result in much memory being saved through
partial reads) and drop bmp_buf_pos altogether, with the effect of greatly
simplifying the bitmap positioning code.