Added UEFI driver inode caching
ntfs-3g is very explicit that when using caching (which is enabled by
default), an application should never attempt to open the same inode
twice. This however does not sit too well with UEFI, and especially
the UEFI Shell, that often does things like opening 2 simultaneous
handles to the same file.

To alleviate potential issues, we add a cache mechanism (linked list)
to keep track of all EFI_NTFS_FILE for which we have an inode. From
there, we make use of lookup calls to prevent inodes from being re-
opened. We also use this lookup to ensure that the parent inode is
closed on NtfsCloseFile(), as required for inode synchronization.