Added UEFI driver rw/ro access mode support
Internally, there are really two access modes that a file can be opened
with for UEFI: Either read/write or read-only.

As such, it doesn't make much sense to multiply file handle structures,
just to know if the file was opened for read-only access or or not. A
more elegant solution is to use the file handle address itself to tell
us that, by duplicating the EFI_FILE part of an EFI_NTFS_FILE struct
(which UEFI forces us to always return as the start of a file handle)
and check which of the two copies has been used as the handle pointer.

Whereas this may seem wasteful, much of this wastage is being recovered
by being able to reuse a single structure for all the file handles to a
specific file, which ultimately helps reduce code complexity.